requirements.txt
#base
Django==1.8.3
Pillow==2.9

#templates
#=======
django-widget-tweaks==1.4.1

#slug
django-autoslug
